+++ lib/isodoc/headlesshtml_convert.rb
@@ -1,12 +1,10 @@
 require_relative "html_function/comments"
-require_relative "html_function/footnotes"
 require_relative "html_function/html"
 require "fileutils"
 
 module IsoDoc
   class HeadlessHtmlConvert < ::IsoDoc::Convert
     include HtmlFunction::Comments
-    include HtmlFunction::Footnotes
     include HtmlFunction::Html
 
     def tmpimagedir_suffix
