+++ lib/isodoc/html_convert.rb
@@ -1,5 +1,4 @@
 require_relative "html_function/comments"
-require_relative "html_function/footnotes"
 require_relative "html_function/html"
 require_relative "html_function/postprocess"
 require_relative "html_function/form"
@@ -8,7 +7,6 @@ module IsoDoc
   class HtmlConvert < ::IsoDoc::Convert
 
     include HtmlFunction::Comments
-    include HtmlFunction::Footnotes
     include HtmlFunction::Form
     include HtmlFunction::Html
